requirements.txt
# Requirements automatically generated by pigar.
# https://github.com/damnever/pigar

# collectioncode\collect.py: 9
# collectioncode\utils.py: 7
# wae_api.py: 19
configparser == 4.0.2

# wae_api.py: 18
futures == 3.3.0

# tests\test_api.py: 6
mock == 0.7.2

# wae_api.py: 21
pbkdf2 == 1.3

# wae_api.py: 22,23
pycryptodome == 3.9.7

# tests\test_api.py: 2
pytest == 4.6.9

# collectioncode\utils.py: 1
# tests\test_api.py: 3
requests == 2.23.0

# collectioncode\utils.py: 2,9
urllib3 == 1.25.8
